modif pour test frameTest
juste changement de la classe a executer dans le pom pour tester
