+++ front-end/src/main/java/edu/ezip/ing1/pds/front/MainApp.java
@@ -1,0 +1,12 @@
+package edu.ezip.ing1.pds.front;
+
+public class MainApp {
+    public static void main (String [] args){
+
+    //AccueilUC2 accueilUC2=new AccueilUC2();
+
+
+       EcranAcceuil ecranAcceuil=new EcranAcceuil();
+
+    }
+}
